requirements.txt
uvicorn~=0.20.0
fastapi~=0.92.0
sqlalchemy~=2.0.4
python-dotenv~=1.0.0
psycopg2
alembic~=1.9.4
passlib~=1.7.4
python-jose
python-multipart
pydantic~=1.10.5
dash~=2.8.1

plotly~=5.13.1
pandas~=1.5.3
requests~=2.28.2

numpy~=1.23.5
matplotlib~=3.6.3
seaborn~=0.12.2
joblib~=1.2.0
scikit-learn~=1.2.2
imblearn~=0.0